Complete the curriculum: docs, cross-compiler clean build
The path is now complete: 38 lessons, 338 koans, 4 capstone projects.

Documentation brought in line with what actually ships. docs/CURRICULUM.md
now lists every lesson and what it covers; the README's tier summaries carry
real counts instead of "planned"; projects/README.md explains each capstone
and what to read it for. Every internal link resolves.

Warning hygiene, which matters more than it sounds: a learner's first `make`
must be silent, so that any warning they later see is one they caused.

  - GCC raised eight diagnostics Clang did not, all on koans that provoke
    them deliberately: switch fallthrough, snprintf truncation, strncpy not
    terminating, a bit-field overflow, and a recursion that GCC cannot see
    ends in a longjmp. Each is now wrapped in KOAN_DELIBERATE, and the
    GCC-only warning names were added to the macro.
  - The Makefile passed -Wno-reserved-identifier to GCC, which does not know
    the flag and said so on every file. It is now Clang-only.
  - about_time set a volatile variable without reading it; the koan now
    asserts the sum, which also stops the loop being optimised away.
  - The generated koans/ warned where solutions/ did not: the int blank
    compared against a size_t tripped -Wsign-compare. The comparison now
    casts explicitly, which is the conversion the lesson is about anyway.

Verified on Apple Clang 21 and GCC 16:
  make check-all      passes on both
  solutions           0 warnings on both
  unsolved koans      0 warnings on both, and still fail as intended
  capstones           701 checks, 0 failures, clean under ASan+UBSan

+++ koans/01_foundations/about_types.c
@@ -47,10 +47,12 @@ KOAN(integer_sizes_are_ordered_not_fixed)
     KOAN_TRUE(sizeof(int) <= sizeof(long));
     KOAN_TRUE(sizeof(long) <= sizeof(long long));
 
-    /* And the minimum widths the standard does guarantee: */
-    KOAN_TRUE(sizeof(short) * CHAR_BIT >= __);
-    KOAN_TRUE(sizeof(long) * CHAR_BIT >= __);
-    KOAN_TRUE(sizeof(long long) * CHAR_BIT >= __);
+    /* And the minimum widths the standard does guarantee. The cast is not
+     * decoration: sizeof yields size_t, and comparing it against a signed
+     * int is the conversion trap this lesson is about. */
+    KOAN_TRUE(sizeof(short) * CHAR_BIT >= (size_t)__);
+    KOAN_TRUE(sizeof(long) * CHAR_BIT >= (size_t)__);
+    KOAN_TRUE(sizeof(long long) * CHAR_BIT >= (size_t)__);
 }
 
 /*
